Refactor Code To Conform Convention
