+++ src/main/java/com/spring/boot/tutorials/DemoApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.spring.boot.tutorials;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
