+++ src/main/java/com/spring/boot/tutorials/postgresql/Tutorial.java
@@ -1,4 +1,4 @@
-package com.example.demo.postgresql;
+package com.spring.boot.tutorials.postgresql;
 
 import lombok.Getter;
 import lombok.Setter;
