+++ src/test/java/spring/boot/tutorials/DemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package spring.boot.tutorials;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
